Extract DashboardOnboarded component
