+++ src/modules/dashboard/components/dashboardOnboarded/index.ts
@@ -1,0 +1,2 @@
+export type { IDashboardOnboardedProps } from './dashboardOnboarded';
+export { DashboardOnboarded } from './dashboardOnboarded';
